+++ bigtable-dataflow-parent/bigtable-hbase-beam/src/main/java/com/google/cloud/bigtable/beam/coders/HBaseMutationCoder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.bigtable.dataflow.coders;
+package com.google.cloud.bigtable.beam.coders;
 
 import java.io.IOException;
 import java.io.InputStream;
