+++ bigtable-dataflow-parent/bigtable-hbase-beam/src/test/java/com/google/cloud/bigtable/beam/coders/HBaseMutationCoderTest.java
@@ -13,11 +13,12 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.bigtable.dataflow.coders;
+package com.google.cloud.bigtable.beam.coders;
 
 import static org.apache.hadoop.hbase.util.Bytes.toBytes;
 
 import com.google.bigtable.repackaged.com.google.api.client.util.Clock;
+import com.google.cloud.bigtable.beam.coders.HBaseMutationCoder;
 import com.google.cloud.bigtable.hbase.adapters.PutAdapterUtil;
 
 import java.io.IOException;
