+++ bigtable-dataflow-parent/bigtable-hbase-beam/src/test/java/com/google/cloud/bigtable/hbase/adapters/PutAdapterUtil.java
@@ -16,7 +16,7 @@
 package com.google.cloud.bigtable.hbase.adapters;
 
 import com.google.bigtable.repackaged.com.google.api.client.util.Clock;
-import com.google.cloud.bigtable.dataflow.coders.HBaseMutationCoderTest;
+import com.google.cloud.bigtable.beam.coders.HBaseMutationCoderTest;
 import com.google.cloud.bigtable.hbase.adapters.PutAdapter;
 
 /**
